refactor: type files naming convention updated
