+++ src/lib/package/utils/image/index.d.ts
