+++ src/lib/package/utils/token/index.d.ts
@@ -1,3 +1,3 @@
-import type {Color} from '../range/index.type';
+import type {Color} from '../range';
 
 export type TokenColors = {light: Color; dark?: Color};
